requirements.txt
absl-py==0.7.1
astor==0.8.0
atari-py==0.1.15
attrs==19.1.0
backcall==0.1.0
bitarray==0.9.3
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
coloredlogs==10.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
getch==1.0
google-pasta==0.1.7
grpcio==1.22.0
gym==0.12.5
h5py==2.9.0
humanfriendly==4.18
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
minerl==0.2.3
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psutil==5.6.3
ptyprocess==0.6.0
pydot==1.4.1
pyglet==1.3.2
Pygments==2.4.2
pyparsing==2.4.0
Pyro4==4.76
pyrsistent==0.15.3
pySmartDL==1.3.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
serpent==1.28
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
typing==3.6.6
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
